fix:data insertion and card number generation

+++ lib/DataStore/Account/Account.h
@@ -19,7 +19,7 @@ public:
     Account(unsigned int uid, std::string name);
 
     // from strings
-    explicit Account(const std::vector<std::string> &strings) : Account(std::stoi(strings[0]), strings[1]) {}
+    explicit Account(const std::vector<std::string> &strings) : Account(std::stoul(strings[0]), strings[1]) {}
 
     // to string
     [[nodiscard]] std::string to_string() const;
